changed default dtypes to float32 and int32

+++ roger/core/utilities.py
@@ -7,19 +7,21 @@ def _get_row_no(arr1d, i, size=1, fill_value=0):
     arr = npx.full((npx.size(arr1d),), dtype=bool, fill_value=False)
     arr = update(
         arr,
-        at[:], arr1d == i,
+        at[:],
+        arr1d == i,
     )
     cond = npx.any(arr)
     if where(arr, size=size, fill_value=fill_value)[0].size > 0:
         val = where(arr, size=size, fill_value=fill_value)[0][0]
     else:
         val = 0
-    row_no = npx.int64(npx.where(cond, val, 0))
+    row_no = npx.int32(npx.where(cond, val, 0))
 
     arr_row_no = npx.full((1,), dtype=int, fill_value=False)
     arr_row_no = update(
         arr_row_no,
-        at[:], row_no,
+        at[:],
+        row_no,
     )
 
     return arr_row_no
@@ -29,19 +31,21 @@ def _get_first_row_no(arr1d, i, size=1, fill_value=0):
     arr = npx.full((npx.size(arr1d),), dtype=bool, fill_value=False)
     arr = update(
         arr,
-        at[:], arr1d == i,
+        at[:],
+        arr1d == i,
     )
     cond = npx.any(arr)
     if where(arr, size=size, fill_value=fill_value)[0].size > 0:
         val = where(arr, size=size, fill_value=fill_value)[0][0]
     else:
         val = 0
-    row_no = npx.int64(npx.where(cond, val, 0))
+    row_no = npx.int32(npx.where(cond, val, 0))
 
     arr_row_no = npx.full((1,), dtype=int, fill_value=False)
     arr_row_no = update(
         arr_row_no,
-        at[:], row_no,
+        at[:],
+        row_no,
     )
 
     return arr_row_no
@@ -51,19 +55,21 @@ def _get_last_row_no(arr1d, i, size=1, fill_value=0):
     arr = npx.full((npx.size(arr1d),), dtype=bool, fill_value=False)
     arr = update(
         arr,
-        at[:], arr1d == i,
+        at[:],
+        arr1d == i,
     )
     cond = npx.any(arr)
     if where(arr, size=size, fill_value=fill_value)[0].size > 0:
         val = where(arr, size=size, fill_value=fill_value)[0][-1]
     else:
         val = 0
-    row_no = npx.int64(npx.where(cond, val, 0))
+    row_no = npx.int32(npx.where(cond, val, 0))
 
     arr_row_no = npx.full((1,), dtype=int, fill_value=False)
     arr_row_no = update(
         arr_row_no,
-        at[:], row_no,
+        at[:],
+        row_no,
     )
 
     return arr_row_no
@@ -106,9 +112,10 @@ def create_catch_masks(ks, nz):
 
 def linear_regression_numpy(x, y, params):
     import numpy as np
+
     for i in range(params.shape[0]):
         for j in range(params.shape[1]):
-            w = np.where(x[i, j, :] > 0, 1/np.max(x[i, j, :]), 0)
+            w = np.where(x[i, j, :] > 0, 1 / np.max(x[i, j, :]), 0)
             coeff_intcpt = np.polyfit(x[i, j, :], y[i, j, :], 1, w=w)
             params = update(params, at[i, j, 0], coeff_intcpt[0])
             params = update(params, at[i, j, 1], coeff_intcpt[1])
@@ -118,6 +125,7 @@ def linear_regression_numpy(x, y, params):
 
 def linear_regression_jax(x, y, params):
     import jax
+
     LEARNING_RATE = 0.005
 
     def init(params):
@@ -130,7 +138,7 @@ def linear_regression_jax(x, y, params):
 
     def loss(params, xs, ys):
         """Computes the least squares error of the model's predictions on x against y."""
-        weights = jax.numpy.where(xs > 0, 1/jax.numpy.max(xs, axis=-1), jax.numpy.NaN)
+        weights = jax.numpy.where(xs > 0, 1 / jax.numpy.max(xs, axis=-1), jax.numpy.NaN)
         pred = params[:, :, 0, jax.numpy.newaxis] * xs * weights + params[:, :, 1, jax.numpy.newaxis]
         return jax.numpy.nanmean((pred - ys) ** 2, axis=-1)
 
@@ -138,7 +146,9 @@ def linear_regression_jax(x, y, params):
     def update(params, xs, ys):
         """Performs one SGD update step on params using the given data."""
         grad = jax.grad(loss)(params, xs, ys)
-        params = update(params, at[:, :, :], jax.tree_multimap(lambda param, g: param - g * LEARNING_RATE, params, grad))
+        params = update(
+            params, at[:, :, :], jax.tree_multimap(lambda param, g: param - g * LEARNING_RATE, params, grad)
+        )
 
         return params
